refactor: generic component ObjectComponent for groups & events, route eventSettings
