+++ agir/front/views.py
@@ -160,6 +160,9 @@ class FullSupportGroupView(SoftLoginRequiredMixin, ReactSingleObjectView):
 class GroupSettingsView(ReactBaseView):
     bundle_name = "front/app"
 
+class EventSettingsView(ReactBaseView):
+    bundle_name = "front/app"
+
 class AgendaView(ReactBaseView):
     bundle_name = "front/app"
