+++ agir/groups/components/groupPage/GroupSettings/styledComponents.js
@@ -1,0 +1,11 @@
+import React from "react";
+import styled from "styled-components";
+
+import style from "@agir/front/genericComponents/_variables.scss";
+
+export const StyledTitle = styled.h3`
+  width: 100%;
+  margin: 0;
+  margin-top: 2rem;
+  margin-bottom: 0.5rem;
+`;
